requirements.txt
selenium~=4.23.0
pandas~=2.2.2
pyarrow~=17.0.0
boto3~=1.34.146
botocore~=1.34.146
python-dotenv~=1.0.1
urllib3~=2.2.2